DHFPROD-6728: Can now run connected matching and merging steps
Biggest change in here is changing some merging/base functions to pass around content objects instead of documents. That allows access to the URI of each doc without depending on the documents having been loaded from the database.

Also modified jobReport so that it has access to the outputContentArray. Then modified the mastering impl of this so that if the array exists, it's used to get counts instead of queries. 

Added test plumbing so that it's easy to create a project with any number of any kind of steps.